rewrite models with lombok
