+++ src/main/java/com/warehouse/model/warehouse/WarehouseRegion.java
@@ -1,4 +1,4 @@
-package com.warehouse.model.wh;
+package com.warehouse.model.warehouse;
 
 public enum WarehouseRegion {
